Replaced rwlock on keydir with simple mutex.
Fixes: bz://1090

Fixes a beam-wide deadlock due to the keydir RW lock being held
for read in fold_keys while the VM goes in and out of the NIF.
If a merger thread tries to write it will try and write lock
in the put_int call.  If the key folder and writer are on the same
scheduler the put from the merge will block completion of the fold.

The RW locks have been replaced by a mutex and count of key folders.
To write the put_int call must lock the mutex and check there are
not key folders.  If there are it releases the mutex and retries.
Writes may be blocked for a long time if there are many key folders
being started.  A fairer wait/notification system should be implemented
once we are happy this resolves the problem.

There is a stub function bitcask_nifs:put_retries you can trace
on to see the count.