Track oldest tstamp in a bitcask data file
keydir now tracks the oldest tstamp in a given data file so that it's
possible to know when a merge needs to be triggered for expiration
purposes. Expose this info via a new summary_info/1 call. status/1 is
deprecated since it returns a big tuple instead of a tidy record.