Add CRC to hintfile and check during startup.
Calculate CRC of hintfile as it is written and write to
special backward compatible trailer record at the end of
the hints file.

Changed the hint folding code to return {error, blah} more
reliably and added a new fold_keys variant that will
switch to scanning the cask files if there were any issues
with the hint file.