Patching bitcask_nifs for PULSE usage
Erlang messages are sent from the C-code (bitcask_nifs.c). This
will not work when scheduled by PULSE. Since it is not possible

Patching bitcask_nifs for PULSE usage

Erlang messages are sent from the C-code (bitcask_nifs.c). This
will not work when scheduled by PULSE. Since it is not possible
to send messages to registered processes from C, quite a bit of
machinery is needed to keep track of the current PULSE process
so that messages can be re-routed properly.

Everything is ifdef:ed, so when compiling normally, no code is
added to bitcask.so.

Usage:
bitcask_nifs:set_pulse_pid(Pid) is used to store the current
PULSE pid, after that messages sent from the C code will be
routed properly.
If no call is made to bitcask_nifs:set_pulse_pid is made, the
behavior is unspecified!!