Change behavior on CRC error
The old behavior, return a seemingly-corrupted accumulator,
or the new behavior, throw an error and abort subsequent
processing of this data file, are both bad.  So, we do
something different: we continue.

If the CRC is good, we call the fold function and get a
new accumulator.  If the CRC is bad, we complain and
use the old accumulator.  In either case, we continue
processing (what we hope are valid contents of) the
rest of the file.