Fix races with merge by removing wall-clock timestamp comparisons
To fix a race with merging, and to also get rid of the problems
where our use of `os:timestamp()` where the OS's clock goes backward,
I added a new argument to `bitcask_nifs_keydir_put_int()` that has
a boolean connotation: if true, then this keydir put operation is
from an actual `bitcask:put()` operation and not from a merge or
someone opening a bitcask in read-only mode.  And if true, then the
only reaon why we won't accept it is if new entry is for a `file_id`
that is lower than the known largest.

For each put, for any reason, we keep track of the known largest
`file_id` and store it in `keydir->biggest_file_id`.  If an
`bitcask:put()` uses a file id that's smaller than `biggest_file_id`,
then that means that a merge has started since the last real put.
In this case, we return `already_exists`, and the Erlang code will
wrap & start a new data file, so that the new put will always be
appended a datafile with a larger file_id than the file_id used by
a merge.