Fix for merge file delete & frozen fold races
One the keydir is frozen during a fold, it's possible for a
parallel merge to delete files needed for the fold in a race
with opening them.

The fix is to add the concept of a fold iterator generation
number to the NIF and to expose it via the keydir info call.
It's safe to delete files from a successful merge operation
if there is no fold happening or if the current fold is of
a newer generation than the generation saved when the merge
finishes.

Also added: turning on the setuid file permissions bit of
merge input files that are marked pending for delete.  The
setuid bit is then checked & filtered when appropriate.
Setuid data files are deleted at open time by a `read_write`
caller.