Fix a race between opening & closing
If an open and a close race, it's possible for the opening
process to "not see" the active writing file because it's
locked by the still-closing process and therefore isn't
visible to bitcask:readable_files().

Fix: Close the writing file and release the write lock first.
Then release the keydir.