Model fix for scheduling pathology that causes open -> {error,timeout}
If we manage to get a timeout when opening a database, there's a
pathological scheduling delay that is causing starvation.  Expose
the starvation by not considering {error, timeout} a success.

Use an `open_timeout` that should be large enough to push beyond
any pathological run schedule that QC ought to create.