Fix races inside scan_key_files()
The scan_key_files() function is used by both keydir_open() and
merge1().  In the former case, any file open problem should be
fatal; in the latter case, there is no harm if a file cannot be
opened because of an honest race with the deferred deletion
from an earlier merge.

Added a query to the merge deferred delete server to be able
to poll the length of its queue.

Also fix a couple of unused var compiler warnings in EUnit
tests.