Fix race with QC model and bitcask_lockops:read_lock_data()
There is an honest race between a creator of a lock file
and someone reading it: the reader may see a file of zero bytes.
This race should be fixed with an atomic lock file creation,
that fix should only be needed for the case where a lock creator
crashes immediately after creating the lock file but not writing
the file contents.

Allow the model to detect that the zero-byte-lock-file race
happened and to ignore it.