Fix very elusive GC-triggered VM crash caused by pulse_send()
This fix appears to fix a very vexing, elusive VM crash that
is triggered well after the data-corrupting fact by a major
GC event.

After experimenting many different ways with trying to find
the problem, using valgrind + a specific test case that
managed to always provoke the problem.  It was quite difficult
to find a deterministic counter-example -- almost all were
ones that would succeed most of the time and only fail once
every 20 or 50 or 200 attempts.

Fix: use the same env for all parts of the term that we send
     to the PULSE process.