Add Run-eunit-loop.expect
The Run-eunit-loop.expect script is a work-around for a number of
problems that I ran into when using PULSE adn the bitcask_eqc.erl
test model.  It's a mess and could really use a refactoring, but
it does what I needed it to....

Before identifying the cause of SIGSEGV and "Bad tag" aborts,
I wanted to automatically restart testing if the test failed for
either reason.  Ditto for deadlock and timeout problems within
PULSE itself.  The timeout problems have been quite mysterious,
but as far as I can tell it isn't a problem with the bitcask
code.  I would run rebar with network distributed enabled, then
attach to it via "erl -remsh rebar@localhost" and use Redbug to
watch for calls to bitcask, bitcask_nifs, bitcask_lockops, and
bitcask_fileops when one of the timeouts happened ... and there
were zero calls to those functions.

The most common problem I see, as of this commit, are "Invariant
broken" errors, e.g.,

     Invariant broken, <0.28726.1> did send at {"c_src/bitcask_nifs.c",1718} when 'handle_info.WorkerPid' is supposed to be the only running process!!

If I change pulse_send() so that we send messages directly to
their recipient (instead of the PULSE controller), then I see
PULSE timeout errors instead of invariant errors.  {shrug}

Another PULSE problem that I'd see in the past were things
like this:

    =ERROR REPORT==== 10-Nov-2012::17:43:40 ===
    Unknown command for pulse: loaded
    =ERROR REPORT==== 11-Nov-2012::21:31:43 ===
    Unknown command for pulse: [code_server|{module,pulse_gen_server}]
    =ERROR REPORT==== 4-Nov-2012::21:47:14 ===
    Unknown command for pulse: {11,<<121,48,236,98,2,156,137,137,27,144,44>>}
    Unknown command for pulse: [{root,yield},
                                {root,yield},
                                {root,yield},
                                .......