PULSE: avoid bug/feature of end-of-test detection
The "after 1000" clause and retry that's in the new
wait_for_ready() function does not change the very
synchronous nature of what we want to do: wait for
a 'ready' or 'error' atom.  The after & recursion
are necessary to avoid a bug/feature of PULSE's
detection of the end of test, aka figuring out if
any process is still runnable.

Without this patch, it's possible for PULSE to create
a scenario where a folder will block waiting for a
keydir to unfreeze, then wait until the very end of
test execution before PULSE will allow that folder
proc to be scheduled.  PULSE seems to decide that
no processes are runnable, despite the fact that
a post-test check of the pid's mailbox clearly shows
that there is a 'ready' atom sitting in the mailbox.
The after & recursion allows PULSE to operate
correctly.