Fix Travis CI failure of multiple_merges_during_fold_test
The environment that Travis runs test under unfortunately does not
immediately pick up the expected number of files to merge for the
different merge operations in the multiple_merges_during_fold_test.
This commit changes the test to re-try merging until the expected
result is seen (or eventually have the test timeout).