Fix fold start time
The fold code was creating a timestamp *before* the keydir was frozen
and using that to look at a snapshot of the data. This was hidden by the
fact that the snapshot code would give in and give you a closest match
if no entry found for the given timestamp, even if this violated the
snapshot logic. When I cleaned up that logic, this surfaced.

Also, fixing get to return not found in the case an entry for the given
snapshot time was not found.