Refactor bitcask_fileops:fold_keys function
Refactor bitcask_fileops:fold_keys to remove nested case statements
and improve the clarity. Also change the has_valid_hintfile function
to only check for a valid hintfile and not also check for the
existence of a hintfile. This change has the side effect of preventing
the logging of error messages about invalid hintfiles when the case is
just that no hintfile exists.