Add a Makefile target for pulse tests
Add a Makefile target to run the pulse tests. The goal is to make it
as simple as possible to run these tests in order to increase the
likelihood they are run during the normal course of code reviews and
not simply forgotten. The pulse test modules are moved up directly
into the test directory to allow rebar to find them and the test
module is renamed to bitcask_pulse for easier identification.  Also a
new version of rebar that can properly execute the new Makefile target
is added.