Add new NIF, bitcask_nifs_keydir_global_pending_frozen/0, to assist PULSE test
Bitcask's fold semantics are difficult enough to try to predict,
but when a keydir is actually frozen, the job is even more difficult.
This NIF is added to reliably inspect if the keydir was frozen during
a PULSE test case, and if we find a fold or fold_keys problem, we
let it pass.

The new NIF is also tested by an existing EUnit test,
keydir_wait_pending_test().