Fix iterator generation increment
iter_generation should increment every time the number of iterators
reaches zero. The existing code only does it when that happens AND the
keydir was frozen.  iter_generation is used in the delete worker to
avoid deleting files that are still used by a snapshot present in the
keydir when iterators are used.  So the current code may postpone
deleting files for potentially a long time if the keydir doesn't reach
the point where it needs to freeze.