Fix merge not correcting race on tombstone write
If a merge writes a tombstone, it needs to detect whether a new value
was put for that key at the same time and avoid that tombstone
overriding it on disk.

As part of this, removing use of merge start time or epoch. Merge logic
should see latest entries in the keydir. I'm not sure why it bothered
checking timestamps on files and the merge timestamp. I imagine that was
able to hide a pulse failure in the past, but does not seem correct.