Fix fold file open/delete race
This fixes the race between a fold getting an epoch and opening the
files in the Bitcask directory.  If between those two actions a file
needed for that snapshot is deleted by a merge, the fold will return
incorrect data.

A fold will now fetch the list of files that it needs to open and the
epoch to use to access a snapshot of data with one operation. It will
then try to open the files, and if it fails try again with a new set of
files and a new epoch.

One question remains open: Now if someone deletes a file from under
Bitcask, the fold will fail indefinitely. Before, it would eventually
proceed and return whatever data is still left. Obviously we'll need to
let it do something similar, falling back to returning partial data.